some migrations

+++ database/migrations/2024_04_29_164845_create_categories_table.php.php
@@ -10,20 +10,22 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up(): void
-    {
-        Schema::create('categories', function (Blueprint $table) {
-            $table->id();
-            $table->string('title');
-            $table->string('slug')->unique();
-            $table->text('summary')->nullable();
-            $table->string('photo')->nullable();
-            $table->unsignedBigInteger('parent_cat_id')->nullable();
-            $table->unsignedBigInteger('restaurant_id')->nullable();
-            $table->enum('status',['active','inactive'])->default('inactive');
-            $table->foreign('parent_cat_id')->references('id')->on('major_categories')->onDelete('SET NULL');
-            $table->foreign('restaurant_id')->references('id')->on('restaurants')->onDelete('SET NULL');
-            $table->timestamps();
-        });
+    {   
+        if (!Schema::hasTable('riders')) {
+            Schema::create('categories', function (Blueprint $table) {
+                $table->id();
+                $table->string('title');
+                $table->string('slug')->unique();
+                $table->text('summary')->nullable();
+                $table->string('photo')->nullable();
+                $table->unsignedBigInteger('parent_cat_id')->nullable();
+                $table->unsignedBigInteger('restaurant_id')->nullable();
+                $table->enum('status',['active','inactive'])->default('inactive');
+                $table->foreign('parent_cat_id')->references('id')->on('major_categories')->onDelete('SET NULL');
+                $table->foreign('restaurant_id')->references('id')->on('restaurants')->onDelete('SET NULL');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
